% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{ll_sample}
\alias{ll_sample}
\title{Estimate genotype likelihood for sample.}
\usage{
ll_sample(
  data = testData,
  genotype = c(1, 1),
  dslipmax = 1,
  repli = 1:ncol(data$counts),
  filter = 0,
  param = list(m = m, sratio = sratio, contamination_rate = contamination_rate, mu =
    mu, nu_plus = nu_plus, nu_minus = nu_minus)
)
}
\arguments{
\item{data}{TODO: Describe the format}

\item{dslipmax}{Maximum number of motifs between allele and slippage.}

\item{repli}{Replicates to process.}

\item{filter}{Number of reads threshold.}

\item{param}{list. m = expected allele ratio for heterozygotes, sratio = standard deviation of the ratio for heterozygotes, contamination rate = expected proportion of contamination,
nu_plus = rate slippage +, nu_minus = rate: slippage -.}
}
\description{
This function calculates genotype likelihood for all the replicates of the sample.
}
\examples{
ll_sample()
}
\keyword{genotype}
\keyword{likelihood}
\keyword{microsatellites}
