% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{lltotal}
\alias{lltotal}
\title{Estimate genotype likelihood for each replicate separately.}
\usage{
lltotal(
  data = testData,
  genotype = c(1, 1),
  repli = 1,
  dslipmax = 1,
  filter = 0,
  param = list(m = m, sratio = sratio, lambda = lambda, mu = mu, nu_plus = nu_plus,
    nu_minus = nu_minus)
)
}
\arguments{
\item{data}{TODO: Describe the format}

\item{repli}{Replicates to process.}

\item{dslipmax}{Maximum number of motifs between allele and slipage.}

\item{filter}{Number of reads threshold.}

\item{param}{list. m = expected allele ratio for heterozygotes, sratio = standard deviation of the ratio for heterozygotes, contamination rate = expected proportion of contamination,
nu_plus = rate slippage +, nu_minus = rate: slippage -.}
}
\description{
This function calculates genotype likelihood for all the replicates of the sample.
}
\examples{
ll_sample()
}
\keyword{genotype}
\keyword{likelihood}
\keyword{microsatellites}
